WIP: successfully added Events into Calendar Columns
What?
(1) Bug Fixed: AddZero method - in DateUtils: previously we were not able to extract the payload from our API
because we did not have the correct format for accessing the Object. AddZero allows for us to correctly match the syntax that the API requires to access payload contents

(2) View Page - has the correct useEffect which is called when we render the component - now we call the API directly as the component renders, so calendars will always render with the correct contents from the API

(3) Expanded Calendar Column : Added EventCard render for our Calendar Columns

Up Next:
(1) Height and Position Logic
(2) update API to take in start date - tim and end date-time (potentially will have issues dealing with events that end overnight)
(3) 'Active Event' Statemanagement for selected Event and Selected Date