WIP: Implemented React Hook Form
Why?
- Using a library which has already built out functionality for my use case
- easy implementation, quick rather than having to rethink and rebuild everything from scratch

What?
(1) We are using a lot of styled and nested components for our create event form, it makes sense for us to use a toolkit which enables useContext like features

Blocker:
- currently have issues on:
(A) styling the datepicker and Timepicker (might need to park this for later)
(B) still trying to implement register... on time date pickers
(c) need to create parser/handler function to format the values into 1 string to inject into our db
(d) need to built parser to convert the date and input values, from our local timezone to inject as UTC 0 time