Feat: Change DisplayTime settings - 12/24 hr
What? - Progress?
- previously we were passing in time data as a prop to create the timebar - 12hr time
- now we are able to change our settings - 12/12:00/24 hour time to adjust the view of the sidebar

Why was this made?
- time data in our app will be used to set the information for our time selector, so it made sense to rework how we were storing, using and display time data.
- before we get started on changing and editing our selector/picker component - it made sense for us to adjust the our time data and the generation of this data.
It was on the docket to change the way we converted and managed time - especially with it's relationship to our database.
Currently our database is holding UTC string as the source of truth - YYYY-MM-DDTHH:MM:SSSSZ - we wanted this to remain consistent so that there is a logically track we can follow when converting time data into display formats