WIP: GeneralSelectPicker
What?
(1) Building Label, Input and Ul for SelectPicker
(2) selectPicker is able to take in placeholder data, label, list of data to create options
(3) internally handles all of the state to then push out for a HTTP request or a basic JSON file update

Why?
(1) we want our Timepicker logic to become scalable across multiple datatypes (used for our Settings Pages and scalable for future features and inputs)
(2) We want this new SelectPicker to work seamlessly with React Hook Form