WIP: functionality for parsing 24 HOUR time to Date()
What?
(1) created 24hourtimetoDateObject util function
(2) now whenever we select a time using the selector - it will parse this information and transform it into a Date Object for our payload.
(3) create a datepicker to implement with the time selector

Next Up:
(1) logic for handling if a date is occuring before?
(2) duration tag to add on for endtime