WIP: implementing 'start of the week setting' - DatePicker
what?
(1) currently everything is hard coded with the start day being 'Sunday'
(2) we want our startday to be flexible across the app - this will help us change view size (e.g. 4 days) help us change the settings of our DatePicker
(3) we want to also be able to display the dates before the first day of the month to help us improve navigation of the datepicker